+++ app/src/main/java/com/movies/presentation/favourite/ui/adapter/FavouriteMovieAdapter.kt
@@ -6,7 +6,7 @@ import androidx.recyclerview.widget.ListAdapter
 import com.movies.databinding.MovieItemBinding
 import com.movies.presentation.base.data.model.MovieUIModel
 import com.movies.presentation.base.view_holder.MoviesViewHolder
-import com.movies.presentation.diff_util.DiffUtilCallback
+import com.movies.presentation.utils.DiffUtilCallback
 
 class FavouriteMovieAdapter(
     private val onClickCallback: ((MovieUIModel) -> Unit),
