+++ app/src/main/java/com/movies/presentation/utils/DiffUtil.kt
@@ -1,4 +1,4 @@
-package com.movies.presentation.diff_util
+package com.movies.presentation.utils
 
 import androidx.recyclerview.widget.DiffUtil
 import com.movies.presentation.base.data.model.MovieUIModel
